requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cron_descriptor==2.0.6
cycler==0.12.1
Django==5.2.6
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-debug-toolbar==6.0.0
django-filter==25.1
django-js-asset==3.1.2
django-modeltranslation==0.19.17
django-parler==2.3
django-redis==6.0.0
django-rosetta==0.10.2
django-smart-selects==1.7.2
django-timezone-field==7.1
django-tinymce==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
environs==14.3.0
flatbuffers==25.2.10
fonttools==4.60.0
gunicorn==23.0.0
helpers==0.2.0
idna==3.10
inflection==0.5.1
jax==0.7.2
jaxlib==0.7.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
marshmallow==4.0.1
matplotlib==3.10.6
mediapipe==0.10.14
ml_dtypes==0.5.3
modeltranslation==0.25
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
parler==1.0.1
pillow==11.3.0
polib==1.2.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.0
psycopg2==2.9.10
pycparser==2.23
PyGetWindow==0.0.9
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
PyRect==0.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rosetta==0.3
rpds-py==0.27.1
scipy==1.16.2
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
